next hw 6

+++ session6/src/boards/boards.controller.ts
@@ -8,6 +8,7 @@ import {
   Delete,
   UseGuards,
   Request,
+  ForbiddenException,
 } from '@nestjs/common';
 import { BoardsService } from './boards.service';
 import { Board } from './board.entity';
@@ -16,6 +17,7 @@ import { CreateBoardDto } from './dto/create-board.dto';
 import { UpdateBoardDto } from './dto/update-board.dto';
 import { Roles } from 'src/common/decorator/role.decorator';
 import { RolesGuard } from 'src/common/guard/roles.guard';
+import { request } from 'http';
 
 @Controller('boards')
 export class BoardsController {
@@ -48,14 +50,29 @@ export class BoardsController {
   async update(
     @Param('id') id: number,
     @Body() updateBoardDto: UpdateBoardDto,
+    @Request() req,
   ): Promise<Board> {
+    const board = await this.boardsService.findOne(id);
+
+    // Access the user ID from the `user` object
+    if (board.user.id !== req.user.userId && req.user.role !== 'admin') {
+      throw new ForbiddenException('You are not allowed to update this board');
+    }
+
     return this.boardsService.update(id, updateBoardDto);
   }
 
   @Roles('admin') // Roles 데코레이터 추가
   @UseGuards(AuthGuard('jwt'), RolesGuard) // RolesGuard 적용
   @Delete(':id')
-  remove(@Param('id') id: number): Promise<void> {
+  async remove(@Param('id') id: number, @Request() req): Promise<void> {
+    const board = await this.boardsService.findOne(id);
+
+    // Access the user ID from the `user` object
+    if (board.user.id !== req.user.userId && req.user.role !== 'admin') {
+      throw new ForbiddenException('You are not allowed to delete this board');
+    }
+
     return this.boardsService.remove(id);
   }
 }
